server/requirements.txt
pytesseract==0.3.10
pdf2image==1.16.3
python-docx==0.8.11
Pillow==10.0.0
opencv-python==4.8.1.78